% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Make a Manhattan Plot}
\usage{
plot_manhattan(
  data,
  color1 = "cornflowerblue",
  color2 = "blue4",
  threshold_linecolor = "black",
  plot_name = "manhattan_plot.png"
)
}
\arguments{
\item{data}{an data frame object loaded from load_data function}

\item{color1}{Color 1 of the dots in the Manhattan plot}

\item{color2}{Color 2 of the dots in the Manhattan plot}

\item{threshold_linecolor}{the color of the threshold lines at 1e-5 and 1e-8}

\item{plot_name}{name of the Manhattan plot file}
}
\description{
This function makes a Manhattan plot with the default data frame of GWAS
summary statistics from load_data function
}
