% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{Make a Q-Q Plot}
\usage{
plot_qq(data, plot_name = "qq_plot.png")
}
\arguments{
\item{data}{an data frame object loaded from load_data function}

\item{plot_name}{name of the Q-Q plot file}
}
\description{
This function makes a Q-Q plot with the default data frame of GWAS
summary statistics from load_data function
}
