% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_data.R
\name{threshold_data}
\alias{threshold_data}
\title{Threshold the GWAS Summary Statistics}
\usage{
threshold_data(data, threshold = 1e-05)
}
\arguments{
\item{data}{an data frame object loaded from load_data function}

\item{threshold}{the p-value threshold specified, usually at 1e-5 or 1e-8}
}
\value{
data_filtered, a data frame of the filtered summary statistics
}
\description{
This function thresholds the GWAS summary statistics based on the specified p value threshold
}
